// packages/frontend/src/openapi/dnd-api/model/spellResponseDto.ts
import type { AttackResponseDto } from "./attackResponseDto";
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * DnD
 * The DnD API description
 * OpenAPI spec version: 1.0
 */
import type { ItemTypeSpell } from "./itemTypeSpell";
import type { SpellResponseDtoManaCost } from "./spellResponseDtoManaCost";

export interface SpellResponseDto {
  type: ItemTypeSpell;
  manaCost: SpellResponseDtoManaCost;
  attacks: AttackResponseDto[];
  name: string;
  level: number;
  imgUrl: string;
}