// packages/frontend/src/openapi/dnd-api/model/weaponResponseDto.ts
import type { AttackResponseDto } from "./attackResponseDto";
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * DnD
 * The DnD API description
 * OpenAPI spec version: 1.0
 */
import type { ItemTypeWeapon } from "./itemTypeWeapon";

export interface WeaponResponseDto {
  type: ItemTypeWeapon;
  attacks: AttackResponseDto[];
  name: string;
  level: number;
  imgUrl: string;
}